% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demean.R
\name{demean}
\alias{demean}
\title{De-mean or mean center variables}
\usage{
demean(x)
}
\arguments{
\item{x}{a vector of numbers}
}
\value{
A vector of mean-centered numbers (mean = 0)
}
\description{
demean removes the mean or mean-centers a variable (excludes NA values)
}
\note{
This function is just a wrapper for x - mean(x, na.rm = T)
}
\examples{
demean(1:3)

}
\author{
Hause Lin
}
