% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ezddm.R
\name{ezddm}
\alias{ezddm}
\title{Fit Wagenmaker et al.'s (2007) EZ-diffusion model to single subject's aggregate/averaged data.}
\usage{
ezddm(
  propCorrect,
  rtCorrectVariance_seconds,
  rtCorrectMean_seconds,
  s = 0.1,
  nTrials = NULL
)
}
\arguments{
\item{propCorrect}{proportion correct (apply edge correction if necessary)}

\item{rtCorrectVariance_seconds}{variance of correct reaction times (in seconds)}

\item{rtCorrectMean_seconds}{mean of correct reaction times (in seconds)}

\item{s}{scaling parameter (often set to 0.1 or 1.0)}

\item{nTrials}{number of trials (useful for edge correction) (optional)}
}
\value{
A dataframe with a (boundary parameter), v (drift rate parameter), and Ter (non-decision time parameter)
}
\description{
ezddm fits Wagenmaker et al.'s (2007) EZ-diffusion model for two-choice response time tasks using proportion correct, variance of correct response reaction times, and mean of correct reaction times (in seconds).
}
\examples{
ezddm(.802, .112, .723)
}
\references{
\itemize{
\item Wagenmakers, E. J., van der Maas, H. L., & Grasman, R. P. (2007). An EZ-diffusion model for response time and accuracy. Psychonomic Bulletin & Review, 14(1), 3-22. doi:10.3758/BF03194023 (\url{https://link.springer.com/article/10.3758/BF03194023})
}
}
\author{
Hause Lin
}
