% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hausekeep.R
\docType{package}
\name{hausekeep}
\alias{hausekeep}
\alias{hausekeep-package}
\title{hausekeep: a package with misc housekeeping functions}
\description{
research and housekeeping functions. details to follow...
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hauselin/hausekeep}
  \item Report bugs at \url{https://github.com/hauselin/hausekeep/issues}
}

}
\author{
Hause Lin \email{hauselin@gmail.com|}
}
