% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{logit}
\alias{logit}
\title{Logit function}
\usage{
logit(p)
}
\arguments{
\item{p}{a vector of probability values (0 to 1)}
}
\value{
A vector
}
\description{
Applies logit transformation
}
\note{
Formula: log(p / (1 - p))
\cr Returns NA values if input values are not between 0 and 1.
\cr
See \url{https://hausetutorials.netlify.com/posts/2019-04-13-logistic-regression/#inverse-logit-and-logit-functions} for intro to logit and inverse logit functions.
}
\examples{
logit(seq(-2, 2, length.out = 100))

}
\author{
Hause Lin
}
