% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_inverse.R
\name{logit_inverse}
\alias{logit_inverse}
\title{Logistic (sigmoid or inverse logit) function}
\usage{
logit_inverse(x)
}
\arguments{
\item{x}{a vector of numbers}
}
\value{
A vector ranging from 0 to 1
}
\description{
Applies inverse logit (sigmoid or logistic) transformation
}
\note{
Inverse logit can be expressed as exp(x) / (1 + exp(x)) or 1 / (1 + exp(-x))
\cr
See \url{https://hausetutorials.netlify.com/posts/2019-04-13-logistic-regression/#inverse-logit-and-logit-functions} for intro to logit and inverse logit functions.
}
\examples{
logit_inverse(-5:5)

}
\author{
Hause Lin
}
