% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers_mad.R
\name{outliers_mad}
\alias{outliers_mad}
\title{Identify outliers using robust median absolute deviation approach}
\usage{
outliers_mad(x, threshold = 3.0, replace_outlier_value = NA,
show_mad_values = FALSE, show_outlier_indices = FALSE,
b_constant = 1.4826, digits = 2, debug = FALSE)
}
\arguments{
\item{x}{a vector of numbers}

\item{threshold}{value to use as cutoff (Leys et al. recommend 2.5 or 3.0 as default)}

\item{replace_outlier_value}{if value is an outlier, what to replace it with? NA by default}

\item{show_mad_values}{if TRUE, will show deviation score of each value}

\item{show_outlier_indices}{if TRUE, return index/position of outliers}

\item{b_constant}{a constant linked to the assumption of normality of the data, disregarding the abnormality induced by outliers}

\item{digits}{how many digits to round output to}

\item{debug}{if TRUE, print messages (FALSE by default)}
}
\value{
A vector with outliers identified (default converts outliers to NA)
}
\description{
outliers_mad is used to identify outliers in vectors using Leys et al.'s (2003) median absolute deviation approach.
}
\details{
We can identify and remove outliers in our data by identifying data points that are too extreme—either too many standard deviations (SD) away from the mean or too many median absolute deviations (MAD) away from the median. The SD approach might not be ideal with extreme outliers, whereas the MAD approach is much more robust (for comparison of both approaches, see Leys et al., 2013, Journal of Experimental Social Psychology).

b_constant is usually 1.4826, a constant linked to the assumption of normality of the data, disregarding the abnormality induced by outliers (Rousseeuw & Croux, 1993).
}
\examples{
x <- c(1, 3, 3, 6, 8, 10, 10, 1000, -1000) # 1000 is an outlier
outliers_mad(x)
outliers_mad(x, threshold = 3.0)
outliers_mad(x, threshold = 2.5, replace_outlier_value = -999)
outliers_mad(x, threshold = 1.5, show_outlier_indices = TRUE)
outliers_mad(x, threshold = 1.5, show_mad_values = TRUE)
outliers_mad(x, threshold = 1.5, show_mad_values = TRUE, replace_outlier_value = -88)
}
\references{
\itemize{
\item Leys, C., Ley, C., Klein, O., Bernard, P., & Licata, L. (2013). Detecting outliers: Do not use standard deviation around the mean, use absolute deviation around the median. Journal of Experimental Social Psychology, 49(4), 764-766. doi:10.1016/j.jesp.2013.03.013 (\url{https://www.sciencedirect.com/science/article/pii/S0022103113000668})}
}
\seealso{
\code{\link{outliersZ}}
}
\author{
Hause Lin
}
