% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliersMAD.R
\name{outliersMAD}
\alias{outliersMAD}
\title{Identify outliers using robust median absolute deviation approach}
\usage{
outliersMAD(x, MADCutOff = 3.0, replaceOutliersWith = NA,
showMADValues = FALSE, outlierIndices = FALSE, bConstant = 1.4826, digits = 2)
}
\arguments{
\item{x}{a vector of numbers}

\item{MADCutOff}{value to use as cutoff (Leys e tal. recommend 2.5 or 3.0 as default)}

\item{replaceOutliersWith}{if value is an outlier, what to replace it with? NA by default}

\item{showMADValues}{if TRUE, will show deviation score of each value}

\item{outlierIndices}{return index/position of outlier}

\item{bConstant}{a constant linked to the assumption of normality of the data, disregarding the abnormality induced by outliers}

\item{digits}{how many digits/decimals to round output to}
}
\value{
A vector with outliers identified (default converts outliers to NA)
}
\description{
outliersMAD is used to identify outliers in vectors using Leys et al.'s (2003) median absolute deviation approach.
}
\details{
We can identify and remove outliers in our data by identifying data points that are too extreme—either too many standard deviations (SD) away from the mean or too many median absolute deviations (MAD) away from the median. The SD approach might not be ideal with extreme outliers, whereas the MAD approach is much more robust (for comparison of both approaches, see Leys et al., 2013, Journal of Experimental Social Psychology).
}
\examples{
example <- c(1, 3, 3, 6, 8, 10, 10, 1000, -1000) # 1000 is an outlier
outliersMAD(example)
outliersMAD(example, MADCutOff = 3.0)
outliersMAD(example, MADCutOff = 2.5, replaceOutliersWith = -999)
outliersMAD(example, MADCutOff = 1.5, outlierIndices = TRUE)
outliersMAD(example, MADCutOff = 1.5, showMADValues = TRUE)
outliersMAD(example, MADCutOff = 1.5, showMADValues = TRUE, replaceOutliersWith = -88)
}
\references{
\itemize{
\item Leys, C., Ley, C., Klein, O., Bernard, P., & Licata, L. (2013). Detecting outliers: Do not use standard deviation around the mean, use absolute deviation around the median. Journal of Experimental Social Psychology, 49(4), 764-766. doi:10.1016/j.jesp.2013.03.013 (\url{https://www.sciencedirect.com/science/article/pii/S0022103113000668})}
}
\seealso{
\code{\link{outliersZ}}
}
\author{
Hause Lin
}
