% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliersZ.R
\name{outliersZ}
\alias{outliersZ}
\title{Identify outliers using Z-score cut-off method}
\usage{
outliersZ(x, zCutOff = 1.96, replaceOutliersWith = NA,
  outlierIndices = FALSE, showZValues = FALSE, digits = 2)
}
\arguments{
\item{x}{a vector of numbers}

\item{zCutOff}{value to use as cutoff (1.96 is a common value)}

\item{replaceOutliersWith}{if value is an outlier, what to replace it with? NA by default}

\item{outlierIndices}{return index/position of outlier}

\item{showZValues}{if TRUE, will show z score of each value}

\item{digits}{how many digits/decimals to round output to}
}
\value{
A vector with outliers identified (default converts outliers to NA)
}
\description{
outliersZ is used to identify outliers in vectors using Z-score cut-off
}
\note{
This detection method is not as robust as the median absolute deviation outlier detection method.
}
\examples{
example <- c(1, 3, 3, 6, 8, 10, 10, 1000) # 1000 is an outlier
outliersZ(example)
outliersZ(example, zCutOff = 3.0)
outliersZ(example, zCutOff = 1.0, replaceOutliersWith = -999)
outliersZ(example, zCutOff = 1.0, outlierIndices = TRUE)
outliersZ(example, zCutOff = 1.0, showZValues = TRUE)
}
\seealso{
\code{\link{outliersMAD}}
}
\author{
Hause Lin
}
