% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{seWithin}
\alias{seWithin}
\title{Compute standard errors (within-subjects)}
\usage{
seWithin(data = NULL, measurevar, betweenvars = NULL, withinvars = NULL,
idvar = NULL, na.rm = TRUE, conf.interval = 0.95, showNormed = FALSE)
}
\arguments{
\item{data}{a dataframe}

\item{measurevar}{the name(s) of column(s) that contain the variable to be summariezed}

\item{betweenvars}{a vector containing names of columns that are between-subjects variables}

\item{withinvars}{a vector containing names of columns that are within-subjects variables}

\item{idvar}{the name of the column that identifies each subject}

\item{na.rm}{boolean that indicates whether to ignore NA's}

\item{conf.interval}{confidence interval range}

\item{showNormed}{whether to show noramlized results}
}
\value{
A data.table
}
\description{
seWithin is used to compute the standard error(s) for one or more variables, for one or more groups in a dataframe, handling within-subjects variables by removing inter-subject variability
}
\note{
Code adapted from R cookbook (see references).
}
\examples{
result <- seWithin(data = ChickWeight, measurevar = "weight",
betweenvars = "Diet", withinvars = "Time", idvar = "Chick")

# multiple outcome variables
ChickWeight2 <- ChickWeight
ChickWeight2$weight2 <- ChickWeight2$weight * 100 # create a new outcome variable
result <- seWithin(data = ChickWeight2, measurevar = c("weight", "weight2"),
betweenvars = "Diet", withinvars = "Time", idvar = "Chick")
}
\references{
(\url{http://www.cookbook-r.com/Graphs/Plotting_means_and_error_bars_(ggplot2)/})
}
\seealso{
\code{\link{stderror}}

\code{\link{stderror}}
}
\author{
Hause Lin
}
