% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{stderror}
\alias{stderror}
\title{Compute standard errors (between-subjects)}
\usage{
stderror(data = NULL, measurevar, groupvars = NULL,
na.rm = TRUE, conf.interval = 0.95, toNumeric = TRUE)
}
\arguments{
\item{data}{a dataframe}

\item{measurevar}{the name(s) of column(s) that contain the variable to be summariezed}

\item{groupvars}{a vector containing names of columns that contain grouping variables}

\item{na.rm}{boolean that indicates whether to ignore NA's}

\item{conf.interval}{confidence interval range}

\item{toNumeric}{whether to convert variables/columns to numeric whenever possible}
}
\value{
A dataframe (tibble)
}
\description{
se is used to compute the standard error(s) for one or more variables, for one or more groups in a dataframe.
}
\note{
Code adapted from R cookbook (see references).
}
\examples{
stderror(data = mtcars, measurevar = c("mpg", "disp"), groupvars = c("cyl", "am"))
stderror(data = ChickWeight, measurevar = "weight", groupvars = "Diet")
}
\author{
Hause Lin
}
