% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryh.R
\name{summaryh}
\alias{summaryh}
\title{Generate formatted results and effect sizes for manuscripts}
\usage{
summaryh(model, decimal = 2, showTable = FALSE, tbl_es = FALSE, es = 'r', ...)
}
\arguments{
\item{model}{a fitted model}

\item{decimal}{round output to decimal places}

\item{showTable}{show results in table format (returns list)}

\item{tbl_es}{show other effect sizes computed using es function}

\item{es}{effect size measure to report (for now, it's always r but options will be introduced later)}

\item{...}{further arguments passed to or from other methods}
}
\value{
A datatable or a list of datatables (if showTable = TRUE or tbl_es = TRUE)
}
\description{
After fitting models, use summaryh() in place of summary() to get APA (American Psychological Association) formatted output that also includes effect size estimates for each effect (r effect size). Currently supports lm, glm, aov, anova, lmer, lme, t-test, chisq.test, and cor.test. Unfortunately, this function won't write your entire results section for you (yet).
}
\note{
Cohen's d: 0.20 (small), 0.50 (medium), .80 (large) (Cohen, 1992)
\cr
correlation r: .10 (small), .30 (medium), .50 (large)
\cr
R-squared: R2: .02 (small), .13 (medium), .26 (large)
}
\examples{
summaryh(lm(mpg ~ qsec, mtcars))
summaryh(aov(mpg ~ gear, mtcars))
summaryh(cor.test(mtcars$mpg, mtcars$gear), tbl_es = TRUE)
}
\author{
Hause Lin
}
