% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore.R
\name{zscore}
\alias{zscore}
\title{Compute z-score}
\usage{
zscore(x, n)
}
\arguments{
\item{x}{A vector}

\item{n}{use n or n-1 in the denominator when computing SD (default 0 for n; -1 for n = -1)}
}
\value{
A vector
}
\description{
Computes z-score (standardized scores)
}
\note{
Formula: (x - mean(x)) / SD(x)
}
\examples{
zscore(1:10)
zscore(1:10, n = 0) # default n = 0 (SD is computed using n)
zscore(1:10, n = -1) # n = -1 (SD is computed using n-1)
}
\author{
Hause Lin
}
