% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es.R
\name{es}
\alias{es}
\title{Effect size conversion}
\usage{
es(d = NULL, r = NULL, R2 = NULL, f = NULL,
oddsratio = NULL, logoddsratio = NULL,
auc = NULL, decimal = 3, msg = TRUE)
}
\arguments{
\item{d}{a numeric vector containing Cohen's d effect size(s)}

\item{r}{a numeric vector containing correlation r effect size(s)}

\item{R2}{a numeric vector containing r-squared effect size(s)}

\item{f}{a numeric vector containing Cohen's f effect size(s)}

\item{oddsratio}{a numeric vector containing odds ratio effect size(s)}

\item{logoddsratio}{a numeric vector containing log odds ratio effect size(s)}

\item{auc}{a numeric vector containing area-under-curve effect size(s)}

\item{decimal}{a numeric vector indicating decimal places of output}

\item{msg}{a boolean indicating whether to show input effect size(s)}
}
\value{
A dataframe with converted effect sizes
}
\description{
es is used to convert one effect size to other effect size measures. Currently supports Cohen's d, corelation r, r-squared, Cohen's f, odds ratio, log odds ratio, and area-under-curve (auc). Also available as a Shiny app here: https://hause.shinyapps.io/shinyapp_effectsize/
}
\details{
Formulae for conversion
\cr \cr
\code{f = d / 2}
\cr
\code{r = d / sqrt(d^2 + 4)}
\cr
\code{d = (2 * r) / sqrt(1 - r^2)}
\cr
\code{R2 = r^2}
\cr
\code{oddsratio = exp(d / (sqrt(3) / pi))}
\cr
\code{logoddsratio = d / (sqrt(3) / pi)}
\cr
\code{auc = pnorm(d / sqrt(2), 0, 1)}
\cr
}
\note{
All conversions assume equal-sized groups. Effect size conventions:
\cr
Cohen's d: 0.20 (small), 0.50 (medium), .80 (large) (Cohen, 1992)
\cr
correlation r: .10 (small), .30 (medium), .50 (large)
\cr
R-squared: R2: .02 (small), .13 (medium), .26 (large)
}
\examples{
es(d = 0.3)
es(r = c(0.1, 0.3))
}
\references{
Borenstein, M., Hedges, L. V., Higgins, J. P. T., & Rothstein, H. R. (2009). Introduction to meta-analysis. Chichester, West Sussex, UK: Wiley.
\cr \cr
Cohen, J. (1988). Statistical power analysis for the behavioral sciences (2nd ed.), Hillsdale, NJ: Erlbaum.
\cr \cr
Rosenthal, R. (1994). Parametric measures of effect size. In H. Cooper & L. V. Hedges (Eds.), The Handbook of Research Synthesis. New York, NY: Sage.
\cr \cr
Ruscio, J. (2008). A probability-based measure of effect size: Robustness to base rates and other factors. Psychological Methods, 13(1), 19-30. doi:10.1037/1082-989x.13.1.19
}
\author{
Hause Lin
}
