% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{normWithin}
\alias{normWithin}
\title{Norms the data within specified groups in the dataframe}
\usage{
normWithin(data = NULL, idvar, measurevar, betweenvars = NULL,
  na.rm = TRUE)
}
\arguments{
\item{data}{a dataframe}

\item{idvar}{the name of the column that identifies each subject}

\item{measurevar}{the name of a column that contains the variable to be summariezed}

\item{betweenvars}{a vector containing names of columns that are between-subjects variables}

\item{na.rm}{boolean that indicates whether to ignore NA's}
}
\value{
a data.frame
}
\description{
normWithin norms the data within specified groups in a data frame
}
\examples{
normWithin(data = sleep, idvar = "ID", measurevar = "extra", betweenvars = "group")
}
