% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sca_lm.R
\name{sca_lm}
\alias{sca_lm}
\title{Specification curve analysis (SCA) for linear regression}
\usage{
sca_lm(data, dv, ivs, covariates = NULL)
}
\arguments{
\item{data}{dataset}

\item{dv}{dependent variable (outcome variable) (character)}

\item{ivs}{independent variable(s) or predictor(s) (character)}

\item{covariates}{covariates (character)}
}
\value{
A tibble/data.table with results
}
\description{
sca_lm is used to run specification curve analysis using linear regression. Fits every possible model for your specified dependent variable, predictor variables, and covariates
}
\details{
Details to follow
}
\examples{
sca_lm(data = mtcars, dv = "mpg", ivs = c("cyl", "carb"), covariates = c("disp", "vs"))
sca_lm(data = mtcars, dv = "mpg", ivs = c("cyl", "carb"))
}
\references{
References to follow
\cr \cr
}
\author{
Hause Lin
}
